refactor(ui): Refactor products/$productId components
These include all product statistics cards, plus the repositories table
and all of its column components showing run and job data.
