+++ ui/src/routes/organizations/$orgId/products/$productId/-components/last-run-date.tsx
@@ -17,9 +17,9 @@
  * License-Filename: LICENSE
  */
 
+import { useQuery } from '@tanstack/react-query';
 import { Loader2 } from 'lucide-react';
 
-import { useRepositoriesServiceGetApiV1RepositoriesByRepositoryIdRuns } from '@/api/queries';
 import { TimestampWithUTC } from '@/components/timestamp-with-utc';
 import {
   Tooltip,
@@ -27,29 +27,26 @@ import {
   TooltipTrigger,
 } from '@/components/ui/tooltip';
 import { config } from '@/config';
+import { getOrtRunsByRepositoryIdOptions } from '@/hey-api/@tanstack/react-query.gen';
 
 export const LastRunDate = ({ repoId }: { repoId: number }) => {
   const {
     data: runs,
     isPending: runsIsPending,
     isError: runsIsError,
-  } = useRepositoriesServiceGetApiV1RepositoriesByRepositoryIdRuns(
-    {
-      repositoryId: repoId,
-      limit: 1,
-      sort: '-index',
+  } = useQuery({
+    ...getOrtRunsByRepositoryIdOptions({
+      path: { repositoryId: repoId },
+      query: { limit: 1, sort: '-index' },
+    }),
+    refetchInterval: (query) => {
+      const curData = query.state.data?.data;
+      if (curData && curData[0] && curData[0].finishedAt) {
+        return undefined;
+      }
+      return config.pollInterval;
     },
-    undefined,
-    {
-      refetchInterval: (query) => {
-        const curData = query.state.data?.data;
-        if (curData && curData[0] && curData[0].finishedAt) {
-          return undefined;
-        }
-        return config.pollInterval;
-      },
-    }
-  );
+  });
 
   if (runsIsPending) {
     return (
